+++ views/forms/BarChartFields.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import Delete from '@material-ui/icons/Delete';
 import TextField from '@material-ui/core/TextField'
 import Button from '@material-ui/core/Button'
