+++ views/layouts/AppLayout.jsx
@@ -1,20 +1,30 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect } from 'react'
+import { compose } from 'redux'
+import { connect } from 'react-redux'
+import { withStyles } from '@material-ui/core/styles'
 
-import Header from '../components/Header';
-import Sidebar from '../components/Sidebar';
+import Header from '../components/Header'
+import Sidebar from '../components/Sidebar'
 
-import getAvatar from '../fe-utils/avatarUtil';
-import getCounts from '../fe-utils/countUtil';
+import getAvatar from '../fe-utils/avatarUtil'
+import getCounts from '../fe-utils/countUtil'
 import { fetchSubjects } from '../fe-utils/fetchUtil'
+import getDefaultStyles from '../fe-utils/styleUtil'
 
-const AppLayout = ({ user, title }) => {
+
+const AppLayout = ({
+  user,
+  classes,
+  title,
+  children
+}) => {
   const [openDrawer, setOpenDrawer] = useState(false)
   const [sideBarState, setSideBarState] = useState({ 
     totalDays: 0,
     totalStudies: 0,
     totalSubjects: 0 
   })
-  const [avatar, setAvatar] = useState('');
+  const [avatar, setAvatar] = useState('')
 
   const toggleDrawer = () => setOpenDrawer(!openDrawer)
 
@@ -26,7 +36,7 @@ const AppLayout = ({ user, title }) => {
   }, [])
 
   return (
-    <>
+    <div className={classes.root}>
       <Header
         handleDrawerToggle={toggleDrawer}
         title={title}
@@ -41,8 +51,18 @@ const AppLayout = ({ user, title }) => {
         totalSubjects={sideBarState.totalSubjects}
         user={user}
       />
-    </>
+      <div className={`${classes.content} ${classes.contentPadded}`}>
+        {children}
+      </div>
+    </div>
   )
 }
 
-export default AppLayout
+const styles = theme => ({
+  ...getDefaultStyles(theme),
+})
+const mapStateToProps = (state) => ({
+  user: state.user
+})
+
+export default compose(withStyles(styles, { withTheme: true }), connect(mapStateToProps))(AppLayout)
